Separated admin and user emailer routes. Removed unnessacary route for time being. Added auth checks to respective routes.
